Create FishingRodController.cs
This script assumes that you have already set up the fishing rod and attached it to the player character. It also assumes that you have defined the line start point and end point, as well as the lure or bait object.

The script checks for input from the player to cast and reel in the line. When the player presses the spacebar, the script applies a force to the lure to cast the line. If the player holds down the spacebar while the line is casting, the script moves the lure towards the rod to reel in the line. Once the lure reaches the rod, the script stops reeling in the line and resets the casting and reeling variables.

You can modify this script to fit the specific needs of your game, such as adding tension to the line, creating animations for casting and reeling, or adjusting the casting power and reel speed.